Promote case-fold column resolver to libsq/driver
The cross-source insert case-folding fix in postgres.TableColumnTypes
(da5aa52e) used a private resolveTableColumnsFold helper. ClickHouse
needs the same logic for the same reason — its dialect is case-sensitive
when identifiers are backtick-quoted — so factor the helper out as
driver.ResolveTableColumnsFold to keep both call sites in sync.

The function is pure string/slice logic with no driver-specific
knowledge. Postgres now calls the shared helper; the unit test moves
from drivers/postgres/internal_test.go to
libsq/driver/colmatch_test.go.
